requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
firebase-admin==6.6.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
motor==3.6.0
msgpack==1.1.0
paho-mqtt==2.1.0
passlib==1.7.4
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.9.1
pyparsing==3.2.0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
requests==2.32.3
rsa==4.9
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.38.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
redis==6.2.0
python-dateutil==2.9.0